Clean ui utils and component violations
