+++ frontend/src/metabase/css/core/bordered.module.css
@@ -43,18 +43,6 @@
   border-left: var(--border-size) var(--border-style) var(--mb-color-border);
 }
 
-.borderLight {
-  border-color: color-mix(
-    in srgb,
-    var(--mb-color-border) 20%,
-    transparent
-  ) !important;
-}
-
-.borderError {
-  border-color: var(--mb-color-error) !important;
-}
-
 .borderSuccess {
   border-color: var(--mb-color-success) !important;
 }
@@ -72,8 +60,4 @@
 /* ensure there is no border via important */
 .borderless {
   border: none !important;
-}
-
-.borderMedium {
-  border-width: var(--border-size-medium);
 }
