+++ frontend/src/metabase/css/core/flex.module.css
@@ -16,10 +16,6 @@
   flex: 1 0 auto;
 }
 
-.flexHalf {
-  flex: 0.5;
-}
-
 .flex1 {
   flex: 1;
 }
@@ -80,10 +76,6 @@
   align-items: flex-start;
 }
 
-.alignEnd {
-  align-items: flex-end;
-}
-
 .alignStretch {
   align-items: stretch;
 }
@@ -115,10 +107,6 @@
 
 .flexWrap {
   flex-wrap: wrap;
-}
-
-.flexReverse {
-  flex-direction: row-reverse;
 }
 
 /* Contents of elements inside flex items might not be wrapped correctly on IE11,
