+++ frontend/src/metabase/css/core/grid.module.css
@@ -9,15 +9,3 @@
 .GridCell {
   flex: 1;
 }
-
-.Grid1of2 > .GridCell {
-  flex: 0 0 50%;
-}
-
-.GridGutters {
-  margin: -1em 0 1em -1em;
-}
-
-.GridGutters > .GridCell {
-  padding: 1em 0 0 1em;
-}
