+++ frontend/src/metabase/css/core/hide.module.css
@@ -21,14 +21,3 @@
     display: inherit !important;
   }
 }
-
-.TableInteractiveDetailButton {
-  color: var(--mb-color-brand);
-  transition: opacity 0.2s ease-in-out;
-  margin: 0 4px;
-  opacity: 0;
-}
-
-.show .TableInteractiveDetailButton {
-  opacity: 1;
-}
