+++ frontend/src/metabase/css/core/layout.module.css
@@ -42,10 +42,6 @@
   width: 100%;
 }
 
-.half {
-  width: 50%;
-}
-
 /* position utilities */
 .fixed {
   position: fixed;
