+++ frontend/src/metabase/css/core/text.module.css
@@ -24,21 +24,10 @@
   font-weight: 700;
 }
 
-.textHeavy {
-  font-weight: 900;
-}
-
 /* text style */
 
 .textItalic {
   font-style: italic;
-}
-
-/* larger text size used for descriptions  */
-.textBody {
-  font-size: 1.286em;
-  line-height: 1.457em;
-  color: var(--mb-color-text-secondary); /* TODO - is this bad? */
 }
 
 .textList {
@@ -64,10 +53,6 @@
 
 .textSmall {
   font-size: 0.875em;
-}
-
-.textSmaller {
-  font-size: 0.8em;
 }
 
 .textEllipsis {
@@ -101,8 +86,4 @@
 
 .textPreWrap {
   white-space: pre-wrap;
-}
-
-.breakAnywhere {
-  line-break: anywhere;
 }
