+++ frontend/src/metabase/css/core/width.module.css
@@ -1,7 +1,3 @@
 .wFull {
   width: 100%;
 }
-
-.ParameterDropdownWidth {
-  width: 18.75rem;
-}
