+++ frontend/src/metabase/css/core/z-index.module.css
@@ -16,15 +16,3 @@
 .z3 {
   z-index: 3;
 }
-
-.z4 {
-  z-index: 4;
-}
-
-.z5 {
-  z-index: 5;
-}
-
-.z6 {
-  z-index: 6;
-}
