+++ frontend/src/metabase/lib/types.ts
@@ -1,5 +1,12 @@
 import _ from "underscore";
 
+/**
+ * Makes every property in the object optional, recursively.
+ */
+export type DeepPartial<T> = {
+  [K in keyof T]?: T[K] extends object ? DeepPartial<T[K]> : T[K];
+};
+
 export const isNotNull = <T>(value: T | null | undefined): value is T => {
   return value != null;
 };
