+++ frontend/src/metabase/plugins/oss/collections.ts
@@ -3,6 +3,10 @@ import type { ComponentType } from "react";
 import { t } from "ttag";
 
 import { Messages } from "metabase/admin/permissions/constants/messages";
+import type {
+  CollectionAuthorityLevelConfig,
+  CollectionInstanceAnaltyicsConfig,
+} from "metabase/collections/types";
 import { getIconBase } from "metabase/lib/icon";
 import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
 import type { IconProps } from "metabase/ui";
@@ -10,10 +14,8 @@ import type {
   BaseEntityId,
   Bookmark,
   Collection,
-  CollectionAuthorityLevelConfig,
   CollectionEssentials,
   CollectionId,
-  CollectionInstanceAnaltyicsConfig,
 } from "metabase-types/api";
 
 // Types
