+++ frontend/src/metabase/ui/components/data-display/Ellipsified/Ellipsified.stories.tsx
