+++ frontend/src/metabase/ui/components/data-display/Ellipsified/index.ts
@@ -1,0 +1,1 @@
+export { Ellipsified } from "./Ellipsified";
