+++ frontend/src/metabase/ui/components/inputs/TextareaBlurChange/TextareaBlurChange.tsx
@@ -3,7 +3,7 @@ import type { ChangeEvent, FocusEvent, KeyboardEvent } from "react";
 import { useCallback, useLayoutEffect, useRef, useState } from "react";
 import { flushSync } from "react-dom";
 
-import { useUnmountLayout } from "metabase/common/hooks/use-unmount-layout";
+import { useUnmountLayout } from "metabase/ui/hooks/use-unmount-layout";
 
 import { Textarea } from "../Textarea";
