+++ frontend/src/metabase/ui/index.ts
@@ -14,3 +14,9 @@ export type {
 export { useHover, useElementSize } from "@mantine/hooks";
 
 export * from "./components";
+export type {
+  MetabaseComponentTheme,
+  ChartColor,
+  ColorCssVariableOrString,
+} from "./component-theme";
+export { DEFAULT_METABASE_COMPONENT_THEME } from "./component-theme";
