+++ frontend/src/metabase/visualizations/lib/color-scales.unit.spec.ts
@@ -1,0 +1,26 @@
+import Color from "color";
+
+import { color } from "metabase/ui/colors/palette";
+
+import { getColorScale } from "./color-scales";
+
+describe("getColorScale", () => {
+  const colors = [
+    Color(color("background-primary")).hex(),
+    Color(color("background-primary-inverse")).hex(),
+  ];
+
+  it("should interpolate colors by default", () => {
+    const scale = getColorScale([0, 1], colors);
+
+    expect(scale(0.2).toUpperCase()).not.toEqual(colors[0]);
+    expect(scale(0.8).toUpperCase()).not.toEqual(colors[1]);
+  });
+
+  it("should not interpolate colors when specified", () => {
+    const scale = getColorScale([0, 1], colors, true);
+
+    expect(scale(0.2).toUpperCase()).toEqual(colors[0]);
+    expect(scale(0.8).toUpperCase()).toEqual(colors[1]);
+  });
+});
