+++ frontend/src/metabase/visualizations/lib/table_format.ts
@@ -5,7 +5,7 @@ import Color from "color";
 
 import { isNumber } from "metabase/lib/types";
 import { alpha } from "metabase/ui/colors";
-import { getLinearColorScale, getSafeColor } from "metabase/ui/colors/scales";
+import { getSafeColor } from "metabase/ui/colors/scales";
 import type {
   ColumnFormattingOperator,
   ColumnFormattingSetting,
@@ -15,6 +15,8 @@ import type {
 } from "metabase-types/api";
 
 import type { Extent, Formatter } from "../types";
+
+import { getLinearColorScale } from "./color-scales";
 
 type Formatters = Record<string, Formatter[]>;
